requirements.txt
numpy>=1.24.4
pandas>=2.0.3
PyYAML==6.0.*
scikit-learn>=1.3.0
pytorch>=2.8.0
